refactor(feedback): changing naming to single convention
