+++ src/features/feedback/feedback-instructor/instructor-tickets.tsx
@@ -4,22 +4,24 @@ import Sheet from '@mui/joy/Sheet';
 import Chip from '@mui/joy/Chip';
 import Typography from '@mui/joy/Typography';
 import Skeleton from '@mui/joy/Skeleton';
+import Stack from '@mui/joy/Stack';
 import { ColorPaletteProp } from '@mui/joy';
 
-import { useGetFeedbackTicketsQuery } from 'app/api/feedback.slice';
-
+import { useInstructorTicketsQuery } from 'app/api/feedback.slice';
 import { FeedbackTicket } from 'shared/ts/types';
 import { parseDate } from 'shared/lib/functions';
-import ViewTicketModal from './components/viewTicketModal';
 
-export default function TicketsTable() {
-  const { data, isLoading } = useGetFeedbackTicketsQuery(undefined);
+import ChangeStatusModal from './status-change';
+
+export default function InstructorTickets() {
+  const { data, isLoading } = useInstructorTicketsQuery(undefined);
 
   return (
     <Sheet
       variant="outlined"
       sx={{
         display: "flex",
+        justifyContent: "center",
         flexDirection: "column",
       }}
     >
@@ -102,7 +104,9 @@ export default function TicketsTable() {
                     <Typography level="body-xs">{parseDate(ticket.updated)}</Typography>
                   </td>
                   <td>
-                    <ViewTicketModal {...ticket} />
+                    <Stack direction="row" p={1}>
+                      <ChangeStatusModal {...ticket} />
+                    </Stack>
                   </td>
                 </tr>
               ))}
