+++ src/features/feedback/tickets-table.tsx
@@ -4,24 +4,22 @@ import Sheet from '@mui/joy/Sheet';
 import Chip from '@mui/joy/Chip';
 import Typography from '@mui/joy/Typography';
 import Skeleton from '@mui/joy/Skeleton';
-import Stack from '@mui/joy/Stack';
 import { ColorPaletteProp } from '@mui/joy';
 
-import { useInstructorTicketsQuery } from 'app/api/feedback.slice';
+import { useGetFeedbackTicketsQuery } from 'app/api/feedback.slice';
+
 import { FeedbackTicket } from 'shared/ts/types';
 import { parseDate } from 'shared/lib/functions';
+import ViewTicketModal from './components/view-ticket';
 
-import ChangeStatusModal from './statusChangeModal';
-
-export default function InstructorTickets() {
-  const { data, isLoading } = useInstructorTicketsQuery(undefined);
+export default function TicketsTable() {
+  const { data, isLoading } = useGetFeedbackTicketsQuery(undefined);
 
   return (
     <Sheet
       variant="outlined"
       sx={{
         display: "flex",
-        justifyContent: "center",
         flexDirection: "column",
       }}
     >
@@ -104,9 +102,7 @@ export default function InstructorTickets() {
                     <Typography level="body-xs">{parseDate(ticket.updated)}</Typography>
                   </td>
                   <td>
-                    <Stack direction="row" p={1}>
-                      <ChangeStatusModal {...ticket} />
-                    </Stack>
+                    <ViewTicketModal {...ticket} />
                   </td>
                 </tr>
               ))}
